gooter, refactoring
