+++ src/components/about/aboutContent.js
@@ -1,0 +1,9 @@
+export const aboutContent = {
+  heading: "Educator turned statistician turned developer",
+  overview:
+    "Back end, front end, data engineering, machine learning. I'm a passionate developer in love with problem solving. My present interest is web and app development, but I'm always open to new problems.",
+  current:
+    "Currently I am a React developer and data engineer with Travelers Insurance. My focus is developing apps for consuming and training Natural Language models.",
+  github: "https://github.com/smhaley/",
+  linkedIn: "https://www.linkedin.com/in/shawn-haley/",
+};
