+++ src/components/project/ProjectContainer.js
@@ -5,8 +5,10 @@ import pep from "../../pics/pep.png";
 import tab from "../../pics/tab.png";
 import Project from "./Project";
 import { pepContent, tabContent } from "./content";
+import { Heading } from "../../styles/common/generic";
 
 const ProjContainer = styled.section`
+  margin-top: 100px;
   display: flex;
   flex-wrap: wrap;
   flex-direction: column;
@@ -26,42 +28,6 @@ const ProjContainer = styled.section`
       visibility: hidden;
     `}
 `;
-
-
-const Heading = styled.div`
-  color: ${(props) => props.theme.green};
-  h3 {
-    font-family: Consolas, monospace;
-    font-size: 1rem;
-    margin-bottom: 20px;
-    @media screen and (min-width: ${(props) => props.theme.xs}) {
-      font-size: 1.3rem;
-      &::before,
-      &::after {
-        display: inline-block;
-        content: "";
-        border-top: 0.05rem solid ${(props) => props.theme.blue};
-        width: 2rem;
-        margin: 0 1.5rem;
-        transform: translateY(-0.4rem);
-      }
-    }
-    @media screen and (min-width: ${(props) => props.theme.sm}) {
-      font-size: 1.7rem;
-      &::before,
-      &::after {
-        width: 4rem;
-        margin: 0 1.5rem;
-        transform: translateY(-0.5rem);
-      }
-    }
-  }
-
-  @media screen and (min-width: ${(props) => props.theme.md}) {
-    margin-top: -52px;
-  }
-`;
-
 
 const Stack = styled.div`
   margin-top: 100px;
